feat: Rename project from YTS Movies to Brokeflix and update related configurations
