+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.rayat.dev.ytsmovies">
+    package="dev.rayat.brokeflix">
     <uses-permission android:name="android.permission.INTERNET"/>
 
     <!-- Legacy write access for API 24-29 (Android 7-10). Scoped storage on
@@ -34,7 +34,7 @@
         </intent>
     </queries>
    <application
-        android:label="YTS Movies"
+        android:label="Brokeflix"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher"
         android:allowBackup="false"
