+++ pubspec.yaml
@@ -56,11 +56,11 @@ dependencies:
   tray_manager: ^0.5.2
 
 dev_dependencies:
-  flutter_launcher_icons: ^0.14.3
+  flutter_launcher_icons: ^0.14.4
   flutter_native_splash: ^2.2.19
   change_app_package_name: ^1.1.0
 
-  package_rename: ^1.2.0
+  package_rename: ^1.10.1
   flutter_lints: ^6.0.0
   freezed: ^3.0.6
   json_serializable: ^6.9.0
